+++ migrations/2026-08-04-000001_token_revoked_retention_index/down.sql
@@ -1,0 +1,1 @@
+DROP INDEX CONCURRENTLY idx_tokens_revoked_retention;